requirements.txt
appdirs==1.4.4
asyncio==3.4.3
distlib==0.3.1
filelock==3.0.12
prettytable==2.0.0
PyYAML==5.3.1
six==1.15.0
virtualenv==20.2.2
wcwidth==0.2.5
